pyproject.toml
[project]
name = "openai-chat"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ivan", email = "molotkov.1997@mail.ru" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "pytelegrambotapi>=4.17.0",
    "openai>=1.21.2",
    "psycopg2>=2.9.9",
    "aiohttp>=3.9.5",
    "greenlet>=3.0.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/openai_c.hat"]
